+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/supply/supplyDetail/SuppliesDetailUiState.kt
@@ -10,6 +10,8 @@ data class SuppliesDetailUiState(
     val supplyBatchList: SupplyBatchExt? = null,
     val isLoading: Boolean = false,
     val error: String? = null,
+    // Controls the visibility of the delete confirmation dialog
     val decisionDialogVisible: Boolean = false,
+    // Controls whether the snackbar should be shown
     val snackbarVisible: Boolean = false,
 )